+++ core/src/math/quaternion.cpp
@@ -5,12 +5,12 @@ namespace spruce {
 	}
 
 	quaternion::quaternion(float pitch, float yaw, float roll) {
-		float sy = sin(yaw / 2.0f);
-		float sp = sin(pitch / 2.0f);
-		float sr = sin(roll / 2.0f);
-		float cy = cos(yaw / 2.0f);
-		float cp = cos(pitch / 2.0f);
-		float cr = cos(roll / 2.0f);
+		float sy = std::sin(yaw / 2.0f);
+		float sp = std::sin(pitch / 2.0f);
+		float sr = std::sin(roll / 2.0f);
+		float cy = std::cos(yaw / 2.0f);
+		float cp = std::cos(pitch / 2.0f);
+		float cr = std::cos(roll / 2.0f);
 		x = cy * sp * cr + sy * cp * sr;
 		y = sy * cp * cr - cy * sp * sr;
 		z = cy * cp * sr - sy * sp * cr;
@@ -32,7 +32,7 @@ namespace spruce {
 		this->x = axis.x * sinHalfAngle;
 		this->y = axis.y * sinHalfAngle;
 		this->z = axis.z * sinHalfAngle;
-		this->w = cos(angle / 2.0f);
+		this->w = std::cos(angle / 2.0f);
 		nor();
 	}
 
@@ -57,7 +57,7 @@ namespace spruce {
 		this->x = axis.x * sinHalfAngle;
 		this->y = axis.y * sinHalfAngle;
 		this->z = axis.z * sinHalfAngle;
-		this->w = cos(angle / 2.0f);
+		this->w = std::cos(angle / 2.0f);
 		nor();
 		return *this;
 	}
@@ -73,12 +73,12 @@ namespace spruce {
 	}
 
 	quaternion& quaternion::set(float yaw, float pitch, float roll) {
-		float sy = sin(yaw / 2.0f);
-		float sp = sin(pitch / 2.0f);
-		float sr = sin(roll / 2.0f);
-		float cy = cos(yaw / 2.0f);
-		float cp = cos(pitch / 2.0f);
-		float cr = cos(roll / 2.0f);
+		float sy = std::sin(yaw / 2.0f);
+		float sp = std::sin(pitch / 2.0f);
+		float sr = std::sin(roll / 2.0f);
+		float cy = std::cos(yaw / 2.0f);
+		float cp = std::cos(pitch / 2.0f);
+		float cr = std::cos(roll / 2.0f);
 		x = cy * sp * cr + sy * cp * sr;
 		y = sy * cp * cr - cy * sp * sr;
 		z = cy * cp * sr - sy * sp * cr;
@@ -89,11 +89,11 @@ namespace spruce {
 
 	void quaternion::toEuler(float& yaw, float& pitch, float& roll) {
 		float t = y * x + z * w;
-		int8 pole = t > 0.499f ? 1 : (t < -0.499f ? -1 : 0);
+		float pole = t > 0.499f ? 1.0f : (t < -0.499f ? -1.0f : 0.0f);
 		bool locked = !(pole == 0);
-		yaw = !locked ? atan2(2.0f * (y * w + x * z), 1.0f - 2.0f * (y * y + x * x)) : 0.0f;
-		pitch = !locked ? asin(2.0f * (w * x - z * y)) : pole * static_cast<float>(M_PI) * 0.5f;
-		roll = !locked ? atan2(2.0f * (w * z + y * x), 1.0f - 2.0f * (x * x + z * z)) : pole * 2.0f * atan2(y, w);
+		yaw = !locked ? std::atan2(2.0f * (y * w + x * z), 1.0f - 2.0f * (y * y + x * x)) : 0.0f;
+		pitch = !locked ? std::asin(2.0f * (w * x - z * y)) : pole * static_cast<float>(M_PI) * 0.5f;
+		roll = !locked ? std::atan2(2.0f * (w * z + y * x), 1.0f - 2.0f * (x * x + z * z)) : pole * 2.0f * atan2(y, w);
 	}
 
 	float quaternion::mag2() const {
@@ -101,7 +101,7 @@ namespace spruce {
 	}
 
 	float quaternion::mag() const {
-		return sqrt(mag2());
+		return std::sqrt(mag2());
 	}
 
 	quaternion& quaternion::conjugate() {
